+++ lib/packages/wikibase/data-model/src/Statement/StatementList.php
@@ -17,8 +17,11 @@ use Wikibase\DataModel\Snak\SnakList;
 use Wikibase\DataModel\Snak\Snaks;
 
 /**
- * Ordered, non-unique, mutable, collection of Statement objects.
- * Provides various filter operations though does not do any indexing by default.
+ * Ordered, non-unique, collection of Statement objects.
+ * Provides various filter operations.
+ *
+ * Does not do any indexing by default.
+ * Does not provide complex modification functionality.
  *
  * @since 1.0
  *
@@ -143,6 +146,45 @@ class StatementList implements IteratorAggregate, Comparable, Countable {
 		}
 
 		return new self( $statements );
+	}
+
+	/**
+	 * @since 2.4
+	 *
+	 * @param int|int[] $acceptableRanks
+	 *
+	 * @return StatementList
+	 */
+	public function getWithRank( $acceptableRanks ) {
+		$acceptableRanks = (array)$acceptableRanks;
+		$statements = array();
+
+		foreach ( $this->statements as $statement ) {
+			if ( in_array( $statement->getRank(), $acceptableRanks ) ) {
+				$statements[] = $statement;
+			}
+		}
+
+		return new self( $statements );
+	}
+
+	/**
+	 * Returns the so called "best statements".
+	 * If there are preferred statements, then this is all the preferred statements.
+	 * If there are no preferred statements, then this is all normal statements.
+	 *
+	 * @since 2.4
+	 *
+	 * @return StatementList
+	 */
+	public function getBestStatements() {
+		$statements = $this->getWithRank( Statement::RANK_PREFERRED );
+
+		if ( !$statements->isEmpty() ) {
+			return $statements;
+		}
+
+		return $this->getWithRank( Statement::RANK_NORMAL );
 	}
 
 	/**