+++ lib/packages/wikibase/data-model/tests/unit/Statement/StatementListTest.php
@@ -445,4 +445,86 @@ class StatementListTest extends \PHPUnit_Framework_TestCase {
 		);
 	}
 
+	public function testGivenInvalidRank_getWithRankReturnsEmptyList() {
+		$list = new StatementList();
+		$this->assertEquals( new StatementList(), $list->getWithRank( 42 ) );
+		$this->assertEquals( new StatementList(), $list->getWithRank( 42, 1337 ) );
+	}
+
+	public function testGivenValidRank_getWithRankReturnsOnlyMatchingStatements() {
+		$statement = new Statement( new Claim( new PropertyNoValueSnak( 42 ) ) );
+		$statement->setRank( Statement::RANK_PREFERRED );
+
+		$secondStatement = new Statement( new Claim( new PropertyNoValueSnak( 1337 ) ) );
+		$secondStatement->setRank( Statement::RANK_NORMAL );
+
+		$thirdStatement = new Statement( new Claim( new PropertyNoValueSnak( 9001 ) ) );
+		$thirdStatement->setRank( Statement::RANK_DEPRECATED );
+
+		$list = new StatementList( array( $statement, $secondStatement, $thirdStatement ) );
+
+		$this->assertEquals(
+			new StatementList( array( $statement ) ),
+			$list->getWithRank( Statement::RANK_PREFERRED )
+		);
+
+		$this->assertEquals(
+			new StatementList( array( $secondStatement, $thirdStatement ) ),
+			$list->getWithRank( array( Statement::RANK_NORMAL, Statement::RANK_DEPRECATED ) )
+		);
+	}
+
+	public function testWhenListIsEmpty_getBestStatementsReturnsEmptyList() {
+		$list = new StatementList();
+		$this->assertEquals( new StatementList(), $list->getBestStatements() );
+	}
+
+	public function testWhenOnlyDeprecatedStatements_getBestStatementsReturnsEmptyList() {
+		$statement = new Statement( new Claim( new PropertyNoValueSnak( 42 ) ) );
+		$statement->setRank( Statement::RANK_DEPRECATED );
+
+		$secondStatement = new Statement( new Claim( new PropertyNoValueSnak( 9001 ) ) );
+		$secondStatement->setRank( Statement::RANK_DEPRECATED );
+
+		$list = new StatementList( array( $statement, $secondStatement ) );
+		$this->assertEquals( new StatementList(), $list->getBestStatements() );
+	}
+
+	public function testWhenPreferredStatements_getBestStatementsReturnsOnlyThose() {
+		$statement = new Statement( new Claim( new PropertyNoValueSnak( 42 ) ) );
+		$statement->setRank( Statement::RANK_PREFERRED );
+
+		$secondStatement = new Statement( new Claim( new PropertyNoValueSnak( 1337 ) ) );
+		$secondStatement->setRank( Statement::RANK_NORMAL );
+
+		$thirdStatement = new Statement( new Claim( new PropertyNoValueSnak( 9001 ) ) );
+		$thirdStatement->setRank( Statement::RANK_DEPRECATED );
+
+		$fourthStatement = new Statement( new Claim( new PropertyNoValueSnak( 23 ) ) );
+		$fourthStatement->setRank( Statement::RANK_PREFERRED );
+
+		$list = new StatementList( array( $statement, $secondStatement, $thirdStatement, $fourthStatement ) );
+		$this->assertEquals(
+			new StatementList( array( $statement, $fourthStatement ) ),
+			$list->getBestStatements()
+		);
+	}
+
+	public function testWhenNoPreferredStatements_getBestStatementsReturnsOnlyNormalOnes() {
+		$statement = new Statement( new Claim( new PropertyNoValueSnak( 42 ) ) );
+		$statement->setRank( Statement::RANK_NORMAL );
+
+		$secondStatement = new Statement( new Claim( new PropertyNoValueSnak( 1337 ) ) );
+		$secondStatement->setRank( Statement::RANK_NORMAL );
+
+		$thirdStatement = new Statement( new Claim( new PropertyNoValueSnak( 9001 ) ) );
+		$thirdStatement->setRank( Statement::RANK_DEPRECATED );
+
+		$list = new StatementList( array( $statement, $secondStatement, $thirdStatement ) );
+		$this->assertEquals(
+			new StatementList( array( $statement, $secondStatement ) ),
+			$list->getBestStatements()
+		);
+	}
+
 }